# src/backend/TestPlanService/Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:5000",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/projects/{projectId}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_DeleteProject",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestSuites"
        ],
        "operationId": "TestSuites_MoveSuite",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "moveSuiteRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MoveSuiteRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/tests/{testId}": {
      "get": {
        "tags": [
          "TestCases"
        ],
        "operationId": "TestCases_GetTestCase",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "testId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetTestCaseResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TestCases"
        ],
        "operationId": "TestCases_UpdateTest",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "testId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddOrUpdateTestCaseRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TestCases"
        ],
        "operationId": "TestCases_DeleteTest",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "testId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/test/plans/{planId}/suites/{suiteId}/points": {
      "get": {
        "tags": [
          "TestPoints"
        ],
        "operationId": "TestPoints_GetSuiteTestPoints",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "suiteId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "boolean",
            "name": "withCildrenSuites",
            "in": "query",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "planId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetSuiteTestPointsResponse"
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/plans/{planId}/testers": {
      "post": {
        "tags": [
          "TestPoints"
        ],
        "operationId": "TestPoints_SetTester",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetTesterRequest"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "planId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/plans/list/assignusers": {
      "post": {
        "tags": [
          "TestPoints"
        ],
        "operationId": "TestPoints_GetUserListForAssign",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserIdentity"
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/plans/{planId}/suites/{suiteId}/tests/result": {
      "post": {
        "tags": [
          "TestPoints"
        ],
        "operationId": "TestPoints_SetOutcome",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetOutcomeRequest"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "planId",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "suiteId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/projects/{projectId}/test/plans/{planId}/suites/{suiteId}/tests/comment": {
      "post": {
        "tags": [
          "TestPoints"
        ],
        "operationId": "TestPoints_SetComment",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetDescriptionRequest"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "planId",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "suiteId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserIdentity": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "GetTestCaseResponse": {
      "type": "object",
      "required": [
        "id",
        "priority",
        "state",
        "automationStatus"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "priority": {
          "type": "integer",
          "format": "int32"
        },
        "state": {
          "$ref": "#/definitions/WiState"
        },
        "assignedTo": {
          "$ref": "#/definitions/UserIdentity"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "automationStatus": {
          "$ref": "#/definitions/AutomationStatus"
        },
        "automationTestName": {
          "type": "string"
        },
        "automationTestStorage": {
          "type": "string"
        },
        "automationTestType": {
          "type": "string"
        },
        "precondition": {
          "type": "string"
        },
        "postcondition": {
          "type": "string"
        },
        "changeBy": {
          "$ref": "#/definitions/UserIdentity"
        },
        "steps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TestStepInfo"
          }
        }
      }
    },
    "WiState": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Design",
        "Closed",
        "Ready"
      ],
      "enum": [
        0,
        1,
        2
      ]
    },
    "AutomationStatus": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Manual",
        "Automated"
      ],
      "enum": [
        0,
        1
      ]
    },
    "TestStepInfo": {
      "type": "object",
      "required": [
        "id",
        "order"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        },
        "action": {
          "type": "string"
        },
        "result": {
          "type": "string"
        }
      }
    },
...
    "AddOrUpdateTestCaseRequest": {
      "type": "object",
      "required": [
        "priority",
        "state",
        "assignedTo",
        "automationStatus"
      ],
      "properties": {
        "priority": {
          "type": "integer",
          "format": "int32"
        },
        "state": {
          "$ref": "#/definitions/WiState"
        },
        "assignedTo": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "automationStatus": {
          "$ref": "#/definitions/AutomationStatus"
        },
        "automationTestName": {
          "type": "string"
        },
        "automationTestStorage": {
          "type": "string"
        },
        "automationTestType": {
          "type": "string"
        },
        "precondition": {
          "type": "string"
        },
        "postcondition": {
          "type": "string"
        },
        "steps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AddOrUpdateTestStep"
          }
        }
      }
    },
    "AddOrUpdateTestStep": {
      "type": "object",
      "required": [
        "order"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        },
        "action": {
          "type": "string"
        },
        "result": {
          "type": "string"
        }
      }
    },
...
    "GetSuiteTestPointsResponse": {
      "type": "object",
      "properties": {
        "points": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SuiteTestPointItem"
          }
        }
      }
    },
    "SuiteTestPointItem": {
      "type": "object",
      "required": [
        "id",
        "testSuiteId",
        "testConfigId",
        "testCaseId",
        "order",
        "priority",
        "outcome"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "testSuiteId": {
          "type": "integer",
          "format": "int32"
        },
        "testConfigId": {
          "type": "integer",
          "format": "int32"
        },
        "testCaseId": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "order": {
          "type": "integer",
          "format": "int32"
        },
        "priority": {
          "type": "integer",
          "format": "int32"
        },
        "outcome": {
          "$ref": "#/definitions/Outcome"
        },
        "configuration": {
          "type": "string"
        },
        "tester": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "Outcome": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Unknow",
        "Planed",
        "Passed",
        "Failed",
        "Blocked",
        "Skipped",
        "Paused"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ]
    },
    "SetTesterRequest": {
      "type": "object",
      "required": [
        "testerId"
      ],
      "properties": {
        "testPointIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "testerId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "SetOutcomeRequest": {
      "type": "object",
      "required": [
        "outcome"
      ],
      "properties": {
        "testPointIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "outcome": {
          "$ref": "#/definitions/Outcome"
        }
      }
    },
    "SetDescriptionRequest": {
      "type": "object",
      "properties": {
        "testPointIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "MoveSuiteRequest": {
      "type": "object",
      "required": [
        "newParentId"
      ],
      "properties": {
        "newParentId": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}